chore(#5176): register the release authority in the CI lane and ratchet gates
Gate wiring for the three fix commits, kept separate so the behavioural diff
stays readable.

- justfile + tests/test_fast_check_ci_wiring.py: put
  `services::discord::zombie_foreground_release::tests` in the curated
  `test-non-pg` lane. Both directions of the release guard (free the zombie /
  never touch a live turn) now run on every main push instead of sitting outside
  every lane filter.
- The three new PG tests are named `..._pg_...` rather than `..._pg` so the
  pg-less lane's `_pg_` skip actually matches them. Their four pre-existing
  siblings in the same module end in `_pg`, are therefore selected by a lane
  with no Postgres, and sit in the rule2 debt baseline; the new tests add zero
  debt instead of inheriting that shape.
- scripts/lib_test_inventory_manifest.txt, scripts/pg_test_lane_manifest.txt:
  re-pinned. The diff is exactly the 13 new tests.
- ARCHITECTURE.md, docs/generated/route-inventory.md: regenerated.
- The `await_holding_lock` ratchet is untouched: the new fixtures isolate their
  runtime root with `TestEnvVarGuard`, which owns the shared test-env mutex
  inside the guard rather than parking a bare `MutexGuard` across the awaits, so
  no test needs an `#[allow]` and the count stays at its 54 baseline.
- scripts/audit_maintainability_giant_baseline.toml +
  scripts/ratchet_admission_history.toml: `src/services/discord/mod.rs`
  2716 -> 2717 for the one-line `pub(crate) mod zombie_foreground_release;`
  registration. Same shape as the #4486 `bot_role` admission — a module
  declaration cannot be net-zeroed without deleting unrelated content. The +20
  the first draft added was brought back down to that floor: the inflight
  presence helper moved into the new child module and the re-export block was
  dropped in favour of direct paths, so the giant carries the declaration only.
