+++ scripts/lib_test_inventory_manifest.txt
@@ -4557,6 +4557,7 @@ services::discord::tmux::restored_turn_injected_anchor_tests::restored_turn_carr
 services::discord::tmux::streaming_rollover_frozen_prefix_persistence_tests::frozen_prefix_persists_across_iteration_and_restore_for_terminal_delete
 services::discord::tmux::terminal_finalize_liveness_tests::terminal_finalize_stop_candidate_requires_all_terminal_flags
 services::discord::tmux::terminal_finalize_liveness_tests::terminal_finalize_stop_decision_requires_dead_tmux
+services::discord::tmux::tmux_kill_policy::recent_turn_stop_range_tests::death_classification_keeps_the_offset_boundary_without_a_generation
 services::discord::tmux::tmux_kill_policy::recent_turn_stop_range_tests::watcher_range_offset_match_uses_generation_and_offset_without_short_ttl
 services::discord::tmux::tmux_kill_policy::recent_turn_stop_range_tests::watcher_range_sessionless_stop_fails_closed
 services::discord::tmux::tmux_kill_policy::recent_turn_stop_range_tests::watcher_range_stop_without_generation_fails_closed
@@ -6000,6 +6001,14 @@ services::discord::watcher_lifecycle_decision::tests::normal_completion_waits_fo
 services::discord::watcher_lifecycle_decision::tests::stream_disconnect_with_dead_tmux_becomes_stop_signal
 services::discord::watcher_lifecycle_decision::tests::watcher_activity_heartbeat_falls_back_to_now_without_runtime_output
 services::discord::watcher_lifecycle_decision::tests::watcher_activity_heartbeat_uses_runtime_output_mtime
+services::discord::zombie_foreground_release::tests::fixtures::cancelled_turn_with_live_inflight_keeps_its_foreground_slot
+services::discord::zombie_foreground_release::tests::fixtures::uncancelled_live_turn_keeps_its_foreground_slot
+services::discord::zombie_foreground_release::tests::fixtures::zombie_foreground_turn_is_released_and_the_queue_survives
+services::discord::zombie_foreground_release::tests::idle_mailbox_is_a_hold_not_a_release
+services::discord::zombie_foreground_release::tests::issue_5176_zombie_shape_releases
+services::discord::zombie_foreground_release::tests::live_turn_is_never_released_by_any_single_missing_evidence
+services::discord::zombie_foreground_release::tests::only_terminal_tmux_readiness_counts_as_structurally_idle
+services::discord::zombie_foreground_release::tests::verdict_labels_are_stable_for_operator_logs
 services::discord_config_audit::db_agent_drift_tests::db_agent_drift_reports_mismatch_before_sync
 services::discord_config_audit::db_agent_drift_tests::post_sync_clean_drift_does_not_persist_pre_sync_warning
 services::discord_config_audit::db_agent_drift_tests::post_sync_verification_failure_keeps_storage_without_stale_advisory
@@ -6878,7 +6887,10 @@ services::settings::tests::settings_write_response_serializes_ok_contract
 services::settings::tests::startup_pg_runtime_config_writer_keeps_blob_authority_without_mirror_reinsertion
 services::slo::avg_latency_tests::avg_latency_query_excludes_dispatch_noise_sources
 services::slo::avg_latency_tests::noisy_metrics_have_alerts_disabled
+services::stale_turn_reconciler::tests::keyed_reconcile_pg_never_idles_a_live_busy_session
+services::stale_turn_reconciler::tests::keyed_reconcile_pg_releases_busy_session_without_inflight
 services::stale_turn_reconciler::tests::keyed_unchanged_outcome_keeps_live_row_turn_active_pg
+services::stale_turn_reconciler::tests::periodic_sweep_pg_still_requires_an_expired_heartbeat
 services::stale_turn_reconciler::tests::preserved_live_tmux_evidence_keeps_stale_row_busy_pg
 services::stale_turn_reconciler::tests::probe_failure_spinner_and_provider_mismatch_preserve_rows_pg
 services::stale_turn_reconciler::tests::stale_busy_candidates_reconcile_only_after_terminal_liveness_pg
@@ -7135,6 +7147,7 @@ services::turn_cancel_finalizer::tests::text_stop_finalizer_uses_stable_surface_
 services::turn_lifecycle::policy_observability_tests::cancel_observability_emits_unknown_noop_direct_fallback
 services::turn_lifecycle::policy_observability_tests::cleanup_policy_observability_surface_matches_cancel_contract
 services::turn_lifecycle::policy_observability_tests::compute_queue_preserved_detects_disk_and_memory_loss
+services::turn_lifecycle::policy_observability_tests::dropped_queue_message_ids_names_the_lost_user_instructions
 services::turn_orchestrator::active_turn_kind_tests::background_start_yields_to_queued_backlog
 services::turn_orchestrator::active_turn_kind_tests::background_turn_is_active_but_not_blocking
 services::turn_orchestrator::active_turn_kind_tests::background_yields_during_dequeue_to_claim_window
