+++ scripts/pg_test_lane_manifest.txt
@@ -575,7 +575,10 @@ services::settings::tests::metadata_less_put_pg_runtime_config_preserves_explici
 services::settings::tests::put_pg_runtime_config_object_and_non_object_keep_only_blob_authority
 services::settings::tests::runtime_config_pg_delete_failure_rolls_back_blob_and_every_mirror
 services::settings::tests::startup_pg_runtime_config_writer_keeps_blob_authority_without_mirror_reinsertion
+services::stale_turn_reconciler::tests::keyed_reconcile_pg_never_idles_a_live_busy_session
+services::stale_turn_reconciler::tests::keyed_reconcile_pg_releases_busy_session_without_inflight
 services::stale_turn_reconciler::tests::keyed_unchanged_outcome_keeps_live_row_turn_active_pg
+services::stale_turn_reconciler::tests::periodic_sweep_pg_still_requires_an_expired_heartbeat
 services::stale_turn_reconciler::tests::preserved_live_tmux_evidence_keeps_stale_row_busy_pg
 services::stale_turn_reconciler::tests::probe_failure_spinner_and_provider_mismatch_preserve_rows_pg
 services::stale_turn_reconciler::tests::stale_busy_candidates_reconcile_only_after_terminal_liveness_pg
